Move JavaCC classes to common package
